backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
PyYAML==6.0.1
langchain==0.0.350
openai==1.3.7
python-multipart==0.0.6
beautifulsoup4==4.12.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7